Move full reconnect time write into an action instead of suppressing the lint rule
